refactor: OpenAPI 문서 설정 스크립트 분리
